+++ quotebot/src/main/java/com/andmark/quotebot/service/impl/QuoteServiceImpl.java
@@ -229,7 +229,6 @@ public class QuoteServiceImpl implements QuoteService {
             log.debug("quote service: handlePostponeChoiceResponse user input = 'случайно'");
             // run logic of random publication
             scheduleRandomPendingPublication();
-            BotAttributes.setUserCurrentBotState(adminChatId, BotState.START);
         }
         // selection for a specific date and time
         else {
@@ -237,6 +236,7 @@ public class QuoteServiceImpl implements QuoteService {
             try {
                 LocalDateTime pendingTime = LocalDateTime.parse(userInput, DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
                 // send the quote to be saved in the database
+                log.debug("sent quote with pendingTime = {} to be saved in database", pendingTime);
                 postponePublishing(pendingTime);
             } catch (DateTimeParseException e) {
                 // Handle invalid date format error
@@ -249,19 +249,24 @@ public class QuoteServiceImpl implements QuoteService {
     void scheduleRandomPendingPublication() {
         log.debug("random pending publication");
         AvailableDaysResponseDTO availableDays = apiService.findAvailableDays();
-        if (availableDays.getAvailableDay() != null && availableDays.getAvailableDay().isAfter(LocalDateTime.now())) {
-            // The availableDay field is not null and contains a date-time value in the future
+        LocalDateTime availableDay = availableDays.getAvailableDay();
+        String message = availableDays.getMessage();
 
-        } else if (availableDays.getAvailableDay() != null) {
-            // The availableDay field is not null and contains a valid past date-time value
-
+        if (availableDay != null) {
+            // The availableDay field is not null
+            log.debug("availableDay is LocalDateTime");
+            postponePublishing(availableDay);
         } else {
             // The availableDay field is null or doesn't contain a valid date-time value
-
+            log.debug("availableDay is null");
+            BotAttributes.setUserCurrentBotState(adminChatId, BotState.POSTPONE);
+            log.debug("set state = {} to user id = {}", BotAttributes.getUserCurrentBotState(adminChatId), adminChatId);
         }
+        telegramBot.sendMessage(adminChatId, null, message);
     }
 
     void postponePublishing(LocalDateTime pendingTime) {
+        log.debug("postponePublishing with pendingTime = {}", pendingTime);
         QuoteDTO quoteDTO = new QuoteDTO();
         quoteDTO.setId(botAttributes.getQuoteId());
         quoteDTO.setContent(botAttributes.getConfirmedContent());
@@ -336,7 +341,8 @@ public class QuoteServiceImpl implements QuoteService {
         }
         // ask for image choice
         telegramBot.removeKeyboard(chatId);
-        telegramBot.sendMessage(chatId, null, "Выбери изображение, введя цифру в пределах [1-" + botAttributes.getImageUrls().size() + "]");
+        telegramBot.sendMessage(chatId, null, "Выбери изображение, введя цифру в пределах [1-" + botAttributes.getImageUrls().size() + "]"
+                + "\n" + "Или введи [0] для поста без картинки");
         BotAttributes.setUserCurrentBotState(adminChatId, BotState.AWAITING_IMAGE_CHOICE);
     }